Added support for default eval test descriptions
If no description can be read from the YAML file for an eval test,
socrates will build one automatically from the
arguments/input/output/return value of the test.

In addition, the plural() function was moved from prompt.py to
util.py and was used in the review test code.