Added 'edit' mode
This mode allows a coordinator to safely edit a criteria file. Specify the
assignment name and group, and socrates will look in CRITERIA_DIR for
the file, and open a temporary version for editing. If the new criteria file
has no syntax errors or isn't missing any required attributes, the new
criteria file replaces the old one.

It's just like visudo.