Added custom script tests; added safer exit()
Custom Python scripts can now be written for testing Python
files that are too complex to test with just 'eval' tests.

The new exit() function in the "util" module should be
used from now on. It tries to run hooks attached to the
'before_exit' trigger and print any exceptions that are
being handled before exiting.